🦀 chore(Cargo.toml): Add new dependencies 🐛 fix(domain/clip/mod.rs): Add custom error types for Clip domain
To support additional functionality, new dependencies were added to the project's `Cargo.toml` file: `chrono` and `uuid`. `chrono` provides functionality for working with dates and times, while `uuid` provides support for universally unique identifiers.

To better handle errors in the Clip domain, custom error types were added to provide more context for different kinds of errors. The `thiserror` crate was used to create the custom error types, which allows the use of the `#[error("...")]` attribute to provide error messages for each variant of the enum. The `ClipError` enum handles errors related to the `Clip` model, such as invalid passwords, titles, and dates, as well as empty content and parsing errors.

Changes made:
- Added the `chrono`, `thiserror`, and `uuid` dependencies to `Cargo.toml`
- A new `ClipError` enum was added to the `Clip` module in `lib/domain/clip/mod.rs`.
- The `Clip` struct in `lib/domain/clip/mod.rs` now includes a `ClipError` type to handle errors.
Signed-off-by: Sabry Awad <dr.sabry1997@gmail.com>